requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
celery==5.4.0
redis==5.0.0
moviepy==1.0.3
openai-whiser==0.1.0
scenedetect==0.6.1
librosa==0.10.2
boto3==1.35.0
pysrt==1.1.2
anthropic==0.40.0
Pillow==10.4.0
